pyproject.toml
[tool.poetry]
name = "hexlet_code"
version = "0.1.0"
description = ""
authors = ["Akhmatnur Khasanshin"]
packages = [
  { include = "gendiff"},
]

[tool.poetry.dependencies]
python = "^3.8"
pytest-cov = "^2.10.1"
PyYAML = "^5.3.1"
flake8 = "^3.8.4"

[tool.poetry.dev-dependencies]
pytest = "^6.1.0"
flake8 = "^3.8.4"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
